src/requirements.txt
agora-token-builder==1.0.0
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cron-descriptor==1.2.35
Django==4.2
django-agora==0.4
django-autoslug==1.9.9
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-cities-light==3.9.2
django-ckeditor==6.5.1
django-cors-headers==4.2.0
django-debug-toolbar==4.0.0
django-environ==0.10.0
django-filter==23.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-phonenumber-field==7.1.0
django-redis==5.2.0
django-timezone-field==5.0
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
install==1.3.5
kombu==5.2.4
phonenumberslite==8.13.20
Pillow==9.5.0
progressbar2==4.2.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pycparser==2.21
python-crontab==2.7.1
python-dateutil==2.8.2
python-utils==3.5.2
pytz==2023.3
redis==4.5.4
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
Unidecode==1.3.6
vine==5.0.0
wcwidth==0.2.6
zope.event==4.6
zope.interface==6.0
